backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
pypdf
google-cloud-storage
google-cloud-aiplatform
pypdf